fix: keep steered lanes alive through reaper
